generate class from export batch
